Adding Base Code
This code was originally from another repository. It has been moved here and refactored so that it can be run as a stand-alone project, using the compiled code of BharatSim available as a .jar in the lib/ folder. [Compiled BharatSim code pulled and assembled on August 30, 2021.] The model includes the bare IndSciSim model, with two-dose vaccine protection, prevaccinations, and dose-prioritisation. A lockdown may also be implemented. The model can also handle infection through a background FOI.