Updated the compiled BharatSim engine
The new engine no longer has the buggy behaviour that occasionally causes runs to crash stochastically, due to an issue with Scala's TrieMap implementation. It is slightly slower, but runs more consistently. Also, the tests are no longer required for assembly, meaning that we no longer need to comment out the neo4j tests if it is not installed on the our computer.